Improve Haskell implementation, still failing on test 15
